Allow top-level await.
Await can easily be supported at the top level; this change allows it.
This change is designed to make CoffeeScript a more effective platform
for learning to program.  Allowing top-level await allows beginners
to write simple programs that work with asynchronous input before
learning how to use functions.

(This solves a key problem with teaching in Javascript.)